Fix (hopefully) some type errors that occur on 64 bit platforms.

+++ base/dataflow/StaticTests.cpp
@@ -1,5 +1,5 @@
 // Test the static member functions of CRingBuffer.cpp
-
+#include <stdlib.h>
 #include <cppunit/extensions/HelperMacros.h>
 #include <cppunit/Asserter.h>
 #include "Asserts.h"
@@ -157,10 +157,10 @@ void StaticRingTest::format()
   size_t max = pHeader->s_maxConsumer;
 
   EQ(CRingBuffer::getDefaultMaxConsumers(), max);
-  EQ(sizeof(RingHeader), (unsigned int)pHeader->s_producerInfo);
-  EQ(sizeof(RingHeader)+sizeof(ClientInformation), (unsigned int)pHeader->s_firstConsumer);
+  EQ(sizeof(RingHeader), (size_t)pHeader->s_producerInfo);
+  EQ(sizeof(RingHeader)+sizeof(ClientInformation), (size_t)pHeader->s_firstConsumer);
   EQ(sizeof(RingHeader) + (pHeader->s_maxConsumer+1)*sizeof(ClientInformation),
-     (unsigned int)pHeader->s_dataOffset);
+     (size_t)pHeader->s_dataOffset);
   EQ(buf.st_size - pHeader->s_dataOffset, (long int)pHeader->s_dataBytes);
   off_t topoff = pHeader->s_topOffset;
   EQ(buf.st_size -1, topoff);
